added Animal constrcutor in Animal.java and refactored org.example package name to animals
