+++ src/main/java/animals/Animal.java
@@ -1,0 +1,16 @@
+package animals;
+
+public class Animal {
+    private String name;
+    private String type;
+    private String color;
+    private int weight, height;
+
+    public Animal(String name, String type, String color, int height, int weight) {
+        this.name = name;
+        this.type = type;
+        this.color = color;
+        this.height = height;
+        this.weight = weight;
+    }
+}
