+++ src/main/java/animals/Test.java
@@ -1,6 +1,6 @@
-package org.example;
+package animals;
 
-public class Main {
+public class Test {
     public static void main(String[] args) {
         System.out.println("Hello world!");
     }
